requirements.txt
wheel
redis
sqlalchemy[asyncio]
fastapi[all]
botocore
boto3
asyncpg
python-jose[cryptography]
pytest
pydantic
passlib[bcrypt]
psycopg2-binary
pytz
pandas
requests
timezonefinder
gunicorn
motor
pytest-asyncio
aiosqlite
jinja2
pytest-mock
coverage
aiosmtplib
mailchimp3
gevent